server/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
attrs==24.2.0
blivet==3.11.0
blivet-gui==2.6.0
Brlapi==0.8.5
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crypt_r==3.13.1
cryptography==45.0.4
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
distro==1.9.0
dnf==4.23.0
dnspython==2.7.0
email_validator==2.2.0
fedora-third-party==0.10
file-magic==0.4.0
firebase-admin==6.9.0
fros==1.1
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
griffe==1.7.3
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.7
iso639==0.1.4
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.19.1
jsonschema-specifications==2023.11.2
langtable==0.0.68
lazydocs==0.4.8
libcomps==0.1.21
libdnf==0.74.0
libvirt-python==10.6.0
louis==3.33.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.4
mdurl==0.1.2
mod_wsgi==5.0.2
msgpack==1.1.1
nftables==0.1
olefile==0.47
openai==1.86.0
openai-agents==0.0.17
packaging==24.2
pdoc==15.0.4
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.0.0
ply==3.11
productmd==1.45
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.25.1
pycparser==2.20
pycups==2.0.4
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyenchant==3.2.2
Pygments==2.19.1
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.10.1
pykickstart==3.58
pyparsing==3.2.3
pyparted==3.13.0
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5_sip==12.15.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-dotenv==1.1.0
python-linux-procfs==0.7.3
python-meh==0.52
python-multipart==0.0.20
python-pam==2.0.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.1
pyynl @ file:///builddir/build/BUILD/kernel-6.14.9-build/kernel-6.14.9/linux-6.14.9-200.fc41.x86_64/tools/net/ynl
rapidfuzz==3.5.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
rich==14.0.0
rpds-py==0.19.1
rpm==4.20.1
rsa==4.9.1
selinux @ file:///builddir/build/BUILD/libselinux-3.7-build/libselinux-3.7/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.7-build/selinux-3.7/python/sepolicy
setools==4.5.1
setuptools==69.2.0
shellingham==1.5.4
simpleaudio==1.0.4
simpleline==1.9.0
six==1.16.0
sniffio==1.3.1
sos==4.9.1
sse-starlette==2.3.6
starlette==0.47.0
systemd-python==235
tqdm==4.67.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.4.0
uv==0.7.13
uvicorn==0.34.3
xkbregistry==0.3
